Lift common document utility functions to generic location
