+++ backend/app/tests/utils/utils.py
@@ -1,9 +1,12 @@
 import random
 import string
+from uuid import UUID
 
 from fastapi.testclient import TestClient
+from sqlmodel import Session
 
 from app.core.config import settings
+from app.crud.user import get_user_by_email
 
 
 def random_lower_string() -> str:
@@ -24,3 +27,24 @@ def get_superuser_token_headers(client: TestClient) -> dict[str, str]:
     a_token = tokens["access_token"]
     headers = {"Authorization": f"Bearer {a_token}"}
     return headers
+
+
+def get_user_id_by_email(db: Session):
+    user = get_user_by_email(session=db, email=settings.FIRST_SUPERUSER)
+    return user.id
+
+
+class SequentialUuidGenerator:
+    def __init__(self, start=0):
+        self.start = start
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        uu_id = self.peek()
+        self.start += 1
+        return uu_id
+
+    def peek(self):
+        return UUID(int=self.start)
